now is the time to delete the old implementation of rbtree! rename tree2 to tree
